src/deploy/requirements_deploy.txt
boto3==1.35.2
python-dotenv==1.0.1